% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_summary.R
\name{eta_summary}
\alias{eta_summary}
\title{Extension of broom::tidy with effect size and APA formatting.}
\usage{
eta_summary(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further argument to be passed to further methods.}
}
\value{
A tibble.
}
\description{
Extension of broom::tidy with effect size and APA formatting.
}
