% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_to_p_eta.R
\name{f_to_p_eta}
\alias{f_to_p_eta}
\title{Convinience function to convert F value to partial eta squared.}
\usage{
f_to_p_eta(f, df_1 = 1, df_2)
}
\arguments{
\item{f}{Fisher's F value.}

\item{df_1}{Degree of freedom for the effect.}

\item{df_2}{Degree of freedom for the error term.}
}
\value{
A numeric.
}
\description{
Convinience function to convert F value to partial eta squared.
}
