% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_outliers.R
\name{lm_outliers}
\alias{lm_outliers}
\title{Returns a data.frame with a specific model's cook's distance,
 hat values}
\usage{
lm_outliers(data, formula, id)
}
\arguments{
\item{data}{A data.frame.}

\item{formula}{A model formula.}

\item{id}{A column name from \code{data} used to identify observations.}
}
\description{
Returns a data.frame with a specific model's cook's distance,
 hat values
}
\references{
Judd, C. M., McClelland, G. H., & Ryan, C. S. (2009). Data analysis: a model comparison approach (2nd ed). New York ; Hove: Routledge.
}
\author{
Dominique Muller, \email{dominique.muller@univ-grenoble-alpes.fr}
}
