% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negligibility_test.R
\name{negligibility_test}
\alias{negligibility_test}
\title{Negligibility tests.}
\usage{
negligibility_test(formula, term, epsilon, data)
}
\arguments{
\item{formula}{A formula with the model to be tested.}

\item{term}{A string indicating the term to be tested.}

\item{epsilon}{A numeric.}

\item{data}{A data frame.}
}
\value{
An object of class "negligibility_test".
}
\description{
Negligibility tests.
}
