% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_pptx.R
\name{plot_to_pptx}
\alias{plot_to_pptx}
\title{Export a plot to a powerpoint document.}
\usage{
plot_to_pptx(plot_code = ggplot2::last_plot(), file, w_r = 1,
  template = NULL)
}
\arguments{
\item{plot_code}{Plot code. Defalut to ggplot2::last_plot()}

\item{file}{Path where the .pptx file should be saved.}

\item{w_r}{Wide to height ratio. Numeric.}
}
\description{
Export a plot to a powerpoint document.
}
