% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_outliers.R
\name{lm_outliers}
\alias{lm_outliers}
\title{Returns outlier indices}
\usage{
lm_outliers(data, formula, id)
}
\arguments{
\item{data}{A data.frame.}

\item{formula}{A model formula.}

\item{id}{A column name from \code{data} used to identify observations.}
}
\description{
Returns a \code{data.frame} with a specific \code{lm} model's studentized residuals,
residuals' cook's distance, and residuals' hat values.
}
\examples{
library(magrittr)
library(LIPmisc)

mtcars \%T>\%
  lm_outliers(mpg ~ disp) \%>\%
  lm(mpg ~ disp,
     data = .)

}
\references{
Judd, C. M., McClelland, G. H., & Ryan, C. S. (2009). Data analysis: a model comparison approach (2nd ed). New York ; Hove: Routledge.
}
\author{
Dominique Muller, \email{dominique.muller@univ-grenoble-alpes.fr}

Cédric Batailler, \email{cedric.batailler@univ-grenoble-alpes.fr}
}
\keyword{outliers}
